backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
gspread==5.12.0
oauth2client==4.1.3
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.1
openpyxl==3.1.2
PyPDF2==3.0.1
